Barkeep Buffs
This gives the barkeep some better starting gold and the ability to handle most unruly customers if a guard isn't on standby.

Why this is good: The barkeep has long been ignored and doesn't have very much at all to begin with. This lets them set the cook up nicely and handle problem patrons confidently.